requirements.txt
torch
torchvision
pandas
pillow
chromadb
dvc